requirements.txt
Django>=5.2.6
django-taggit>=6.1.0
Pillow>=11.3.0
python-dotenv>=1.1.1
